Equalizer works in linux. Lots of compile warnings from Shibatch stuff, but I cleaned our warnings and made it compileable again (casting issues)... gonna go back to windows to design the FX window better. Great job Si.

+++ Musik/Classes/MusikPlayer.cpp
@@ -275,7 +275,7 @@ void CMusikPlayer::SetFX()
 
 void CMusikPlayer::SetFrequency()
 {
-	FSOUND_SetFrequency( FSOUND_ALL, g_FX.GetFrequency() );
+	FSOUND_SetFrequency( FSOUND_ALL, (int)g_FX.GetFrequency() );
 }
 
 void CMusikPlayer::UpdateUI()
